Removed comment about relative paths
It's actually only occurs when using a symbolic link. Technically it returns the
correct result, though it's a bit unfortunate since a relative path to a symbolic
linked directory would be less error prone if the underlying location the link
was pointing to changed. It's beyond the scope of this function to fix, so the
comment isn't helpful.