Completely overhaul cview class
No longer a template class since it ONLY works on zero-terminated
character arrays. It has a limited number of constructors to minimize
chance of using it on non-zero terminated character arrays. The invalid
constructor from a string_view and the remove_suffix method are marked
as delete so they can't be accidentally used.