Switch from push_back to emplace_back, remove temporary string
Since emplace_back() returns a reference, and back() gives access to the
last string added, it doesn't actually make that much sense to provide
a temporary string. The caller can either get a reference from addEmptyLine
or emplack_back, or they can simply add a blank line and then call back to
get a reference to a string which can be modified.