Remove obsolete code and header files
Fixes #147

With this commit, all code and header files that should no longer be used
or maintained has been removed. Not everything is done -- for example,
the ttFileDlg class needs an overhaul, but it will be modified in place
rather than completely tossed like most of the other older classes.