Add -DUNICODE compiler flag
Fixes #143

I ran this through the multi-project build suite and everything still builds
with one exception where comdlg32.lib needed to be added for the clang
linker to work. If there are bugs caused by this change, they are not showing
up just build compilining and linking multiple projects, so will have to
be handled on a case by case basis.