Switch to a template for dlgListView::SetCurSel
The index gets send to the Windows API as a WPARAM, but this allows
the caller to use int, size_t, LRESULT, or any other variable and still
have it compile without warnings.

Note the (WPARAM) c-style cast -- this lets the compiler determine the
best cast to use based on the caller's parameter type. Otherwise, some
types will generate warnings because a different type of cast is needed.