Add _wx suffix to functions that need const wxString&
You can't use the same name for a function when the parameter differs between "std::string_view" and "const wxString&" because if you then call the function with "foo" the compiler won't know which function to use since both can convert a quoted string. Adding the _wx suffix makes it possible to use quoted strings without confusing the compiler.